move code to /src
