+++ src/graphql.js
@@ -3,7 +3,6 @@ const { makeExecutableSchema } = require('graphql-tools');
 const { typeDefs, resolvers } = require('./schemas');
 
 exports.server = (app) => {
-
   const schema = makeExecutableSchema({ typeDefs, resolvers });
   return graphqlExpress({
     schema,
@@ -12,7 +11,6 @@ exports.server = (app) => {
       knex: app.get('knex')
     }
   });
-
 };
 
 exports.graphiql = graphiqlExpress({ endpointURL: '/graphql' });
