+++ src/models/subscriber.js
@@ -8,16 +8,16 @@ const formatSubscriberData = (subscriberData) => ({
 });
 
 const subscriberModel = {
-  findSubscriber({id, email}, {knex}) {
+  findSubscriber ({id, email}, {knex}) {
     if (!Boolean(id || email)) throw new Error('A valid id or email is required to find a subscriber');
     const where = id ? { subscriber_id: id } : { email_address: email };
     return knex.select('*').from('subscribers').where(where)
     .then(subscriberData => {
-      if (!subscriberData.length) throw new Error(`No subscriber found for ${id ? 'id: ' + id : 'email: ' + email }`);
+      if (!subscriberData.length) throw new Error(`No subscriber found for ${id ? 'id: ' + id : 'email: ' + email}`);
       return formatSubscriberData(subscriberData[0]);
     });
   },
-  findSubscribers({ids, isActive, emails}, {knex}) {
+  findSubscribers ({ids, isActive, emails}, {knex}) {
     return knex.select('*').from('subscribers').modify(queryBuilder => {
       if (ids) queryBuilder.whereIn('subscriber_id', ids);
       if (isActive !== undefined) isActive ? queryBuilder.whereNull('unsubscribed_at') : queryBuilder.whereNotNull('unsubscribed_at');
@@ -28,7 +28,7 @@ const subscriberModel = {
       return subscribersData.map(subscriberData => formatSubscriberData(subscriberData));
     });
   },
-  subscribe(email, {knex}) {
+  subscribe (email, {knex}) {
     if (!validator.isEmail(email)) throw new Error(`Cannot create subscriber with invalid email address: ${email}`);
     return knex.insert({ email_address: email }).into('subscribers').returning('*')
     .then(subscriberData => formatSubscriberData(subscriberData[0]))
@@ -39,11 +39,11 @@ const subscriberModel = {
       throw error;
     });
   },
-  unsubscribe(id, {knex}) {
+  unsubscribe (id, {knex}) {
     return knex('subscribers').update({unsubscribed_at: knex.fn.now()}).where({ subscriber_id: id }).returning('*')
     .then(subscriberData => {
       if (!subscriberData.length) throw new Error(`No subscriber found for id: ${id}`);
-      return formatSubscriberData(subscriberData[0])
+      return formatSubscriberData(subscriberData[0]);
     });
   }
 };
