+++ src/schemas/index.js
@@ -25,21 +25,21 @@ const rootSchema = `
 
 const rootResolvers = {
   Query: {
-    edition(root, { id, publishDate }, context) {
+    edition (root, { id, publishDate }, context) {
       return model.findEdition({ id, publishDate }, context);
     },
-    subscriber(root, { id, email }, context) {
+    subscriber (root, { id, email }, context) {
       return model.findSubscriber({ id, email }, context);
     },
-    subscribers(root, { ids, isActive, emails }, context) {
+    subscribers (root, { ids, isActive, emails }, context) {
       return model.findSubscribers({ isActive, ids, emails }, context);
     }
   },
   Mutation: {
-    subscribe(root, { email }, context) {
+    subscribe (root, { email }, context) {
       return model.subscribe(email, context);
     },
-    unsubscribe(root, { id }, context) {
+    unsubscribe (root, { id }, context) {
       return model.unsubscribe(id, context);
     }
   }
