+++ src/utils/cron.js
@@ -5,13 +5,11 @@ const model = require('../models');
 const send = require('./send');
 
 module.exports = (app) => {
-
   const context = { knex: app.get('knex') };
 
   return new CronJob({
     cronTime: '00 30 5 * * 1-5',
     onTick: () => {
-
       // Runs every weekday (Monday through Friday) at 5:30:00 AM.
       const today = moment().format('YYYY-MM-DD');
       const editionPromise = model.findEdition({ publishDate: today }, context);
@@ -24,7 +22,6 @@ module.exports = (app) => {
         subject: edition.subject,
         html
       }));
-
     },
     start: true, // Start the job right now
     timeZone: 'America/New_York' // Time zone of this job.
